[KZN-3415] extracts all types to types file
